requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
idna==3.10
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
